[feature-12] Set up slices for semester and auth
