--- src/redux/interface.ts
@@ -1,4 +1,0 @@
-export interface semesterState {
-    startDate: Date; 
-    endDate: Date;
-}
