+++ src/store/interface.ts
@@ -1,0 +1,9 @@
+export interface semesterState {
+    startDate: Date; 
+    endDate: Date;
+}
+
+export interface AuthState {
+    username: string | null;
+    isAuthenticated: boolean;
+}
